refactor(structure): :art: Moves Math folder into Core folder
