+++ .vscode/settings.json
@@ -11,6 +11,7 @@
         ]
     },
     "conventionalCommits.scopes": [
-        "roadmap"
+        "roadmap",
+        "structure"
     ]
 }
