+++ tests/unit/Math/Vector2Tests.cpp
@@ -1,4 +1,4 @@
-#include <ASGE/Math/Vector2.hpp>
+#include <ASGE/Core/Math/Vector2.hpp>
 
 #include <gtest/gtest.h>
